further testing, cleaning and refactoring for the outliers triple index
